requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
langchain_google_genai
langchain_openai
langchain_chroma
chromadb
faiss-cpu
tiktoken
PyMuPDF
python-dotenv
argparse
streamlit>=1.23.0
python-dotenv>=0.21.0
asyncio>=3.4.3
langchain>=0.0.267
PyPDF2>=3.0.0
pdf2image>=1.16.3
matplotlib>=3.7.1
numpy>=1.24.3
pandas>=2.0.0
flashrank
reportlab
plotly